- adding our Custom Post Type
Now we add our custom post type and we call it in our plugin init
because then we can also flush the rewrite rules.

NOT flushing the rewrite rules on activation/adding the CPT can mean
that we'll get 404 errors on our site and then we'd expect the user to
know to refrest the permalinks.

Users aren't going to know that.