- check caps and make sure the shortcode works
Now we're checking to make sure that a user can either edit or add a
todo item. If they can they should be able to see our form.

For now we're just returning 'something' to make sure that our shortcode
is in fact functioning.

Remember that shortcodes need to return stuff always or it's going to
end up above you're content and you'll look silly.