- adding a response if you don't have the cap
Now we have a response if you don't have the capability to add items to
the list.

I've also broken out the HTML for the form in to a seperate function. I
did this because now the one function deals with the logic around IF the
user can see the form and my 2nd function deals with the display of the
form.

Trying to stick to the S in SOLID. At least as I understand it.