- capturing the form fields
We define form first because we're going to have to call it a few times
and it's much faster to define it as a var and then call it instead of
calling #form-id .something and then getting the value for it.

As I stepped through each one I refreshed the form and used console.log
to make sure I was getting the information I expected.