- catching our JS with a proper AJAX server side thingy
Now we are catching our AJAX request properly. Notice that WordPress
needs 2 hooks.

1. wp_ajax_*
2. wp_ajax_nopriv_*

The special one is the 2nd one. It runs when a user isn't signed in to a
site. So you really only need it if you need this AJAX action to be
processed by users that aren't logged in to your site.

Simply excluding the nopriv version isn't enough for security though. We
still need to nonce this and check our caps and escape everything. Which
we'll do.

Really this is a minimum function to make sure we wrote decent code. We
used wp_send_json_success so that our console.log statment will return
the success object which is different than the 0 (or failure) we saw
in our last commit.