- noncing the ajax action
Now we leverage the nonce we set in our wp_localize_script function
and pass it to our data var in the JS.

On the server side we use check_ajax_referer. First param is the nonce
we should be checking. 2nd param is where we should be checking for it
in our AJAX values that just came through.

Again minimal code that simply returns a success object up front so we
can tell we wrote sane code.