- now we get some HTML going and get our posts
First we define an HTML wrapper. I' always append '-wrapper' to my first
section so that I have something to easily target your CSS to. Not
having a main wrapper for targeting can mean that you have to resort to
crappy CSS (like !important) to get the styles you want. Having that
wrapper and using it to start all your rules also means that it's
unlikely our CSS wil interfere with other plugins/themes.

We also abstract our 'get' function to get the posts. Sure we could
throw it in the list_tasks function but we're trying to build small
abstracted methods that do one thing.

Our main list_tasks method (function) is assembling our task list which
takes a few steps. Having get_task_posts also means that we have a
single function later that we can leverage to get other types of posts
(like querying for who a task is assigned to).