- getting our task list
Now we have 2 functions to get our tasks.

Our first is the main wrapper for our list of tasks. It checks for a
wp_error on our posts or empty (no posts) so that we can return a sane
message to our users saying that there are no posts/tasks matching their
criteria.

Our second function is abstracting the 'getting' of a single <li> for
our task. That way later we can easily append/add a single task item to
our list.