- refactoring for better long term use
When I 1st started with OOP I would often build components in to a class
like this form we've been using. Then later we'd need it again, and like
an idiot I'd build it again.

Pain ensues.

Now I'd actually start out building forms (and other parts we've already
built) in to a template-taps.php file so that I can call them from any
class that I need them in later.

This refactor moves our forms out of the main class and puts them in a
template tags file.