- refactoring single tasks
Just like before we need the single task function in our shortcode and
our ajax request so we refactor and put it in the template tags.

Now we can handle adding a new task to our lists of tasks.

we do this with our same processing function using the else statement on
the upadated if from earlier.